feat: improve quotes

+++ src/commands/quotes.command.ts
@@ -1,10 +1,11 @@
 import { command } from '@/core/commands'
 import { db } from '@/core/db'
+import { friendlyUID } from '@/utils/string_utils'
 import Discord from 'discord.js'
-import { nanoid } from 'nanoid'
 
 export interface Quote {
-  author: string
+  authorName: string
+  authorUid?: string
   quote: string
   uid: string
   meta?: {
@@ -23,10 +24,10 @@ export default command({
   description: 'Manage quotes',
   examples: [
     '`!quote` - Get a random quote',
-    '`!quote search <query>` - Search for a quote',
-    '`!quote add <quote>` - Add a new quote',
-    '`!quote remove <id>` - Remove quote',
     '`!quote <id>` - Get a specific quote',
+    '`!quote search <query>` - Search for a quote',
+    '`!quote add @author <quote>` - Add a new quote (@author can be a user mention, a plain nickname, or left out)',
+    '`!quote remove <id>` - Remove quote - You can only remove quotes you created to reduce abuse.',
   ],
   async execute(message, args) {
     // Get random quote
@@ -38,9 +39,9 @@ export default command({
     const first = args[0].trim().toLowerCase()
 
     switch (first) {
-      // Search quotes
-      case 'search':
-        searchQuotes(message, args.slice(1))
+      // Add quote
+      case 'add':
+        addNewQuote(message, args.slice(1))
         return
 
       // Remove quote
@@ -48,21 +49,30 @@ export default command({
         removeQuote(message, args[1])
         return
 
-      // Add quote
-      case 'add':
+      // Search quotes
+      case 'search':
+        searchQuotes(message, args.slice(1))
+        return
       default:
         if (first.startsWith('#')) {
-          getSingleQuote(message, args)
+          getSingleQuote(message, args[0].substring(1))
         } else {
-          addNewQuote(message, args.slice(first === 'add' ? 1 : 0))
+          getRandomQuote(message, args)
         }
+        return
     }
   },
 })
 
 const clean = (str: string): string => str.replace(/[\t\n|]+/g, ' ').replace(/\s+/g, ' ')
-const getQuoteStr = (guild: Discord.Guild, { author, quote, uid }: Quote): string =>
-  `"${quote}" - ${author} (${guild.members.cache.get(uid)?.toString() ?? `(ID: #${uid})`})`
+const getQuoteStr = (
+  guild: Discord.Guild,
+  { authorName, authorUid, quote, uid }: Quote,
+): string => {
+  const _authorName = authorUid ? guild.members.cache.get(authorUid)?.displayName : authorName
+  const authorMention = authorUid ? `<@!${authorUid}>` : `@${_authorName}`
+  return `**"${quote}"** - ${authorMention} (ID: #${uid})`
+}
 
 async function getRandomQuote(message: Discord.Message, _args: string[]): Promise<void> {
   const count = await collection.countDocuments()
@@ -107,56 +117,58 @@ async function removeQuote(message: Discord.Message, id: string): Promise<void> 
     message.reply(`Quote ${id} not found`)
     return
   }
-  collection.deleteOne({ uid: id })
-  message.reply(`Quote ${id} deleted - ${getQuoteStr(message.guild!, q)}`)
+  if (q.meta?.createdBy === message.author.id) {
+    await collection.deleteOne({ uid: id })
+    message.reply(`Quote ${id} deleted - ${getQuoteStr(message.guild!, q)}`)
+  } else {
+    message.reply('Oops, you can only remove quotes you created!')
+  }
 }
 
 async function addNewQuote(message: Discord.Message, args: string[]): Promise<void> {
   const [authorRaw, ...restRaw] = args
-  const hasAuthor = /<@!\d+>/.test(authorRaw) || authorRaw.startsWith('@')
-  const author = hasAuthor
-    ? authorRaw.startsWith('@')
-      ? authorRaw.slice(1)
-      : authorRaw
-    : 'Anonymous'
-
-  const differentAuthor = hasAuthor && author !== `<@!${message.author.id}>`
-  const authorName = differentAuthor
-    ? authorRaw.startsWith('@')
-      ? authorRaw.slice(1)
-      : message.mentions.guild!.members.cache.find(
-          (u) => u.user.id === message.mentions.users.first()!.id,
-        )?.displayName
-    : message.member!.displayName
-  const quote = (hasAuthor ? restRaw : [authorRaw, ...restRaw]).join(' ')
+  const hasAuthorUid = /<@\d+>/.test(authorRaw)
+  const authorUid = hasAuthorUid ? authorRaw.slice(2, -1) : undefined
+  const authorAtName = authorRaw.startsWith('@') ? authorRaw.slice(1) : 'Anonymous'
+  const authorName = hasAuthorUid
+    ? message.mentions.members!.get(authorUid!)?.displayName ?? authorAtName
+    : authorAtName
+  const hasAuthor = /<@\d+>/.test(authorRaw) || authorRaw.startsWith('@')
+  const quote = (hasAuthor ? restRaw : [authorRaw, ...restRaw]).join(' ').trim()
+  const differentAuthor = authorUid !== message.author.id
 
   const replies = [
-    `wow! How inspiring. I'll forever remember this${differentAuthor ? `, ${author}` : ''}.`,
-    `are you serious? This is the best quote ever${differentAuthor ? `, ${author}` : ''}!`,
+    `Wow! How inspiring. I'll forever remember this${differentAuthor ? `, ${authorName}` : ''}.`,
+    `Are you serious? This is the best quote ever${differentAuthor ? `, ${authorName}` : ''}!`,
     'OH. MY. GOD. Perfection.',
     'I am putting this on my wall. This is a quote I will hold dear to me always.',
-    `is that real? Woah! Hey,${
-      differentAuthor ? ` ${author},` : ''
+    `Is that real? Woah! Hey,${
+      differentAuthor ? ` ${authorName},` : ''
     } did you ever consider writing a book?! This will sell for millions.`,
-    clean(`okay, this is spooky. I definitely dreamt of ${!hasAuthor ? 'a person' : differentAuthor ? author : 'you'}
-    saying exactly that this week. ${!hasAuthor ? 'Is someone' : differentAuthor ? `Is ${author}` : 'Are you'}
-    prying into my subconscious?`),
-    'consider me floored. If there was an award for amazing quotes, it would be named after this exact one.',
-    'why did no one say this earlier? It HAS to be said!',
+    clean(
+      `Okay, this is spooky. I definitely dreamt of ${!hasAuthor ? 'a person' : differentAuthor ? authorName : 'you'} ` +
+        `saying exactly that this week. ${!hasAuthor ? 'Is someone' : differentAuthor ? `Is ${authorName}` : 'Are you'} prying into my subconscious?`,
+    ),
+    'Consider me floored. If there was an award for amazing quotes, it would be named after this exact one.',
+    'Why did no one say this earlier? It HAS to be said!',
     "I can't believe you withold that quote from me until now. It's way too good to just remain unshared!",
     'I have a pretty large memory capacity for a bot, and I gotta say, I scanned all my other quotes, this one is definitely on the top 10.',
     clean(`Oh, I am DEFINITELY saving this. One day someone will interview me about
     ${
-      !hasAuthor ? 'the best quote I can recall,' : differentAuthor ? author : 'you'
+      !hasAuthor ? 'The best quote I can recall,' : differentAuthor ? authorName + ', ' : 'you,'
     } and I will refer to this moment precisely.`),
-    clean(`you're not serious. Are you serious? You can't be serious. It's impossible there's **this** good a quote just floating around
-    out there. It's probably fictional. Yeah.`),
+    clean(
+      `You're not serious. Are you serious? You can't be serious. It's impossible there's **this** good a quote ` +
+        `just floating around
+    out there. It's probably fictional. Yeah.`,
+    ),
   ]
 
   const quoteObj: Quote = {
     quote,
-    author,
-    uid: nanoid(),
+    authorName,
+    authorUid,
+    uid: friendlyUID(),
     meta: {
       authorCachedName: authorName!,
       createdAt: new Date(),
@@ -170,8 +182,7 @@ async function addNewQuote(message: Discord.Message, args: string[]): Promise<vo
   message.reply(`${replies[Math.floor(Math.random() * replies.length)]}\n${quoteStr}`)
 }
 
-async function getSingleQuote(message: Discord.Message, args: string[]): Promise<void> {
-  const id = args[0].slice(1)
+async function getSingleQuote(message: Discord.Message, id: string): Promise<void> {
   const quote = await collection.findOne({ uid: id })
 
   if (!quote) {
